+++ kgcnn/graph/adapter.py
@@ -2,7 +2,7 @@ import logging
 import numpy as np
 from typing import Union
 
-from kgcnn.utils.adj import get_angle_indices, coordinates_to_distancematrix, invert_distance, \
+from kgcnn.graph.adj import get_angle_indices, coordinates_to_distancematrix, invert_distance, \
     define_adjacency_from_distance, sort_edge_indices, get_angle, add_edges_reverse_indices, \
     rescale_edge_weights_degree_sym, add_self_loops_to_edge_indices, compute_reverse_edges_index_map